@@ -11,7 +11,7 @@ authors = [
 ]
 readme = "README.md"
 license = {text = "MIT"}
-requires-python = ">=3.8"
+requires-python = ">=3.11"
 keywords = ["pii", "privacy", "document", "masking", "presidio", "docling", "data-privacy", "gdpr", "anonymization"]
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -77,58 +62,104 @@ where = ["."]
 include = ["cloakpivot*"]
 exclude = ["tests*"]
 
+# ========================================================================================
+# Tool Configurations - Single Source of Truth
+# ========================================================================================
+
+# ----------------------------------------------------------------------------------------
+# Black - Code Formatter
+# ----------------------------------------------------------------------------------------
 [tool.black]
-line-length = 88
-target-version = ['py39']
+line-length = 100
+target-version = ['py311']
 include = '\.pyi?$'
 extend-exclude = '''
 /(
-  # directories
   \.eggs
   | \.git
   | \.hg
   | \.mypy_cache
   | \.tox
   | \.venv
+  | venv
   | _build
   | buck-out
   | build
   | dist
+  | __pycache__
 )/
 '''
 
+# ----------------------------------------------------------------------------------------
+# Ruff - Fast Python Linter with Integrated isort
+# ----------------------------------------------------------------------------------------
 [tool.ruff]
-line-length = 88
-target-version = "py39"
+line-length = 100
+target-version = "py311"
+exclude = [
+    ".git",
+    ".venv",
+    "venv",
+    "__pycache__",
+    "build",
+    "dist",
+    ".mypy_cache",
+    ".pytest_cache",
+    ".ruff_cache",
+]
 
 [tool.ruff.lint]
 select = [
-    "E",  # pycodestyle errors
-    "W",  # pycodestyle warnings
-    "F",  # pyflakes
-    "I",  # isort
-    "B",  # flake8-bugbear
-    "C4", # flake8-comprehensions
-    "UP", # pyupgrade
+    "E",    # pycodestyle errors
+    "W",    # pycodestyle warnings
+    "F",    # pyflakes
+    "I",    # isort
+    "N",    # pep8-naming
+    "UP",   # pyupgrade
+    "B",    # flake8-bugbear
+    "C4",   # flake8-comprehensions
+    "DTZ",  # flake8-datetimez
+    "T20",  # flake8-print
+    "RET",  # flake8-return
+    "SIM",  # flake8-simplify
+    "PTH",  # flake8-use-pathlib
 ]
 ignore = [
-    "E501",  # line too long, handled by black
+    "E501",  # line too long (handled by black)
     "B008",  # do not perform function calls in argument defaults
-    "C901",  # too complex
+    "T201",  # print found (allow in CLI tools)
+    "T203",  # pprint found (allow for debugging)
 ]
 
 [tool.ruff.lint.per-file-ignores]
-"__init__.py" = ["F401"]
+"__init__.py" = ["F401"]  # unused imports in __init__ files
+"tests/**/*.py" = ["T20"]  # allow print in tests
 
+[tool.ruff.lint.isort]
+known-first-party = ["cloakpivot"]
+section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
+
+# ----------------------------------------------------------------------------------------
+# MyPy - Static Type Checker with Gradual Typing
+# ----------------------------------------------------------------------------------------
 [tool.mypy]
-files = ["cloakpivot/cli"]
-python_version = "3.9"
-warn_return_any = true
+python_version = "3.11"
+files = ["cloakpivot"]
+exclude = [
+    "build",
+    "dist",
+    "tests",
+    "scripts",
+    "venv",
+]
+
+# Gradual typing approach - start permissive, tighten over time
+warn_return_any = false
 warn_unused_configs = true
-disallow_untyped_defs = true
-disallow_incomplete_defs = true
+disallow_untyped_defs = false
+disallow_incomplete_defs = false
 check_untyped_defs = true
-disallow_untyped_decorators = true
+disallow_untyped_decorators = false
 no_implicit_optional = true
 warn_redundant_casts = true
 warn_unused_ignores = true
