@@ -1,19 +1,17 @@
 """Method registration system for adding masking/unmasking methods to DoclingDocument."""
 
-from typing import Optional, List
 import warnings
 
 from docling_core.types import DoclingDocument
 
-from cloakpivot.engine import CloakEngine
 from cloakpivot.core.policies import MaskingPolicy
-
+from cloakpivot.engine import CloakEngine
 
 # Global engine instance for method registration
-_global_engine: Optional[CloakEngine] = None
+_global_engine: CloakEngine | None = None
 
 
-def register_cloak_methods(engine: Optional[CloakEngine] = None) -> None:
+def register_cloak_methods(engine: CloakEngine | None = None) -> None:
     """Register masking/unmasking methods on DoclingDocument.
 
     This allows for natural method chaining:
@@ -45,17 +43,15 @@ def register_cloak_methods(engine: Optional[CloakEngine] = None) -> None:
         _global_engine = CloakEngine()
 
     # Check if methods already registered
-    if hasattr(DoclingDocument, '_cloak_methods_registered'):
+    if hasattr(DoclingDocument, "_cloak_methods_registered"):
         warnings.warn(
             "CloakPivot methods already registered on DoclingDocument. "
             "Re-registering with new engine.",
-            UserWarning
+            UserWarning,
         )
 
     def mask_pii(
-        self,
-        entities: Optional[List[str]] = None,
-        policy: Optional[MaskingPolicy] = None
+        self, entities: list[str] | None = None, policy: MaskingPolicy | None = None
     ):
         """Mask PII in this document and return a CloakedDocument wrapper.
 
@@ -78,16 +74,14 @@ def register_cloak_methods(engine: Optional[CloakEngine] = None) -> None:
             Original DoclingDocument with PII restored
         """
         # Check if this is a CloakedDocument wrapper
-        if hasattr(self, '_cloakmap') and hasattr(self, '_doc'):
+        if hasattr(self, "_cloakmap") and hasattr(self, "_doc"):
             # This is a CloakedDocument wrapper
             return _global_engine.unmask_document(self._doc, self._cloakmap)
-        else:
-            # This is a regular DoclingDocument, return as-is
-            warnings.warn(
-                "unmask_pii() called on non-masked document. Returning document as-is.",
-                UserWarning
-            )
-            return self
+        # This is a regular DoclingDocument, return as-is
+        warnings.warn(
+            "unmask_pii() called on non-masked document. Returning document as-is.", UserWarning
+        )
+        return self
 
     # Register methods on DoclingDocument class
     DoclingDocument.mask_pii = mask_pii
@@ -107,20 +101,20 @@ def unregister_cloak_methods() -> None:
     global _global_engine
 
     # Remove methods if they exist
-    if hasattr(DoclingDocument, 'mask_pii'):
-        delattr(DoclingDocument, 'mask_pii')
-    if hasattr(DoclingDocument, 'unmask_pii'):
-        delattr(DoclingDocument, 'unmask_pii')
-    if hasattr(DoclingDocument, '_cloak_methods_registered'):
-        delattr(DoclingDocument, '_cloak_methods_registered')
-    if hasattr(DoclingDocument, '_cloak_engine'):
-        delattr(DoclingDocument, '_cloak_engine')
+    if hasattr(DoclingDocument, "mask_pii"):
+        delattr(DoclingDocument, "mask_pii")
+    if hasattr(DoclingDocument, "unmask_pii"):
+        delattr(DoclingDocument, "unmask_pii")
+    if hasattr(DoclingDocument, "_cloak_methods_registered"):
+        delattr(DoclingDocument, "_cloak_methods_registered")
+    if hasattr(DoclingDocument, "_cloak_engine"):
+        delattr(DoclingDocument, "_cloak_engine")
 
     # Clear global engine
     _global_engine = None
 
 
-def get_registered_engine() -> Optional[CloakEngine]:
+def get_registered_engine() -> CloakEngine | None:
     """Get the currently registered CloakEngine instance.
 
     Returns:
@@ -135,7 +129,7 @@ def is_registered() -> bool:
     Returns:
         True if methods are registered, False otherwise
     """
-    return hasattr(DoclingDocument, '_cloak_methods_registered')
+    return hasattr(DoclingDocument, "_cloak_methods_registered")
 
 
 def update_engine(engine: CloakEngine) -> None:
@@ -151,8 +145,7 @@ def update_engine(engine: CloakEngine) -> None:
 
     if not is_registered():
         raise RuntimeError(
-            "CloakPivot methods not yet registered. "
-            "Call register_cloak_methods() first."
+            "CloakPivot methods not yet registered. " "Call register_cloak_methods() first."
         )
 
     _global_engine = engine
